Simplify plugin intallation.
Symlinking vendor webroot assets is a pain, and this is 100% guaranteed to work. The plugin
no longer tries to use the chosenhq repository as a vendor, and just includes the assets directly in the
plugin package.

Furthermore, using the vendor scripts from the Chosen plugin is optional. The asset uri path is now configurable
with sensible defaults so the bundled assets work out of the box without any symlinking or other configuration.